part1/requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
kafka-python==1.4.3
requests==2.19.1
schedule==0.5.0
urllib3==1.23
